credential-store/pyproject.toml
[tool.poetry]
name = "credential-store"
version = "0.1.0"
description = "Stores credentials in an encrypted file."

license = "MIT"

authors = [
    "Christoph Kuhnke <christoph.kuhnke@exasol.com>"
]

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0"
sqlcipher3-binary = ">=0.5.0"


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-mock = "^3.7.0"

[tool.pytest.ini_options]
minversion = "6.0"

testpaths = [
    "test"
]
